main.py
"""
USAD - Urban Smart Adaptive Dispatcher
Main entry point for the application
"""

import cv2
import yaml
import logging
import argparse
from pathlib import Path

# Import project modules
from src.vision import VideoProcessor
from src.detection import VehicleDetector, VehicleTracker
from src.signal_control import SignalController, SignalOptimizer
from src.eye_module import ViolationDetector
from src.notification import NotificationManager
from src.logging import EventLogger

def load_config(config_path="config/config.yaml"):
    """
    Load configuration from YAML file
    
    TODO:
    - Read YAML configuration file
    - Validate configuration values
    - Return config dictionary
    """
    pass

def initialize_modules(config):
    """
    Initialize all USAD modules
    
    TODO:
    - Create VideoProcessor instance
    - Create VehicleDetector instance
    - Create VehicleTracker instance
    - Create SignalController instance
    - Create ViolationDetector instance
    - Create NotificationManager instance
    - Create EventLogger instance
    - Return all module instances
    """
    pass

def process_frame(frame, modules, config):
    """
    Process a single video frame through the entire USAD pipeline
    
    TODO:
    1. Preprocess frame (blur, resize if needed)
    2. Vehicle Detection:
       - Background subtraction
       - Contour detection
       - Filter by size
    3. Vehicle Tracking:
       - Match detections with previous frames
       - Assign vehicle IDs
    4. Lane Counting:
       - Count vehicles per lane
    5. Signal Optimization:
       - Calculate new signal timing based on congestion
       - Send command to Arduino
    6. Violation Detection (E.Y.E.):
       - Check for red-light violations
       - Check for yellow-light abuse
       - Check for illegal turns
    7. Logging:
       - Log all events and violations
    8. Visualization (optional):
       - Draw vehicle bounding boxes
       - Draw ROI regions
       - Display signal state
       - Show violation alerts
    
    Return: annotated_frame, detection_results, violations
    """
    pass

def main(args):
    """
    Main application loop
    
    TODO:
    1. Load configuration
    2. Initialize all modules
    3. Open video stream
    4. Main loop:
       - Capture frame
       - Process frame through pipeline
       - Display results
       - Handle user input (ESC to quit)
    5. Cleanup and close resources
    """
    pass

if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="USAD - Urban Smart Adaptive Dispatcher")
    parser.add_argument("--config", default="config/config.yaml", help="Path to config file")
    parser.add_argument("--video", default="0", help="Video source (0 for webcam, or path to file)")
    parser.add_argument("--headless", action="store_true", help="Run without display")
    
    args = parser.parse_args()
    main(args)
