src/licenceplate/plate_recognizer.py
"""
USAD OCR Module
License plate detection and optical character recognition
"""

import cv2
import numpy as np
import pytesseract
import logging
from PIL import Image

class LicensePlateDetector:
    """
    Detects license plates in vehicle regions using contour analysis
    
    TODO:
    - Extract license plate region from vehicle ROI
    - Use edge detection and contour analysis
    - Filter by plate-like dimensions
    - Preprocess for OCR
    """
    
    def __init__(self, config):
        """
        Initialize license plate detector
        
        Args:
            config (dict): Configuration from config.yaml
        """
        self.config = config
        self.ocr_config = config.get('ocr', {})
        
        self.plate_width_range = tuple(self.ocr_config.get('plate_width_range', [40, 200]))
        self.plate_height_range = tuple(self.ocr_config.get('plate_height_range', [20, 100]))
        pass
    
    def detect_plate_region(self, vehicle_roi):
        """
        Detect license plate region within vehicle image
        
        TODO:
        1. Convert to grayscale
        2. Apply edge detection (Canny or Sobel)
        3. Find contours
        4. Filter by aspect ratio (plates are rectangular)
        5. Filter by size (plate_width_range, plate_height_range)
        6. Return bounding box of most likely plate
        
        Args:
            vehicle_roi (np.ndarray): Cropped vehicle region
        
        Returns:
            tuple: (x, y, w, h) of plate region, or None if not found
        """
        pass
    
    def preprocess_for_ocr(self, plate_roi):
        """
        Preprocess plate image for better OCR accuracy
        
        TODO:
        1. Convert to grayscale if needed
        2. Apply thresholding (Otsu or adaptive)
        3. Denoise (bilateral filter or morphological)
        4. Enhance contrast (histogram equalization)
        5. Upscale if too small
        6. Return preprocessed image
        
        Args:
            plate_roi (np.ndarray): License plate region
        
        Returns:
            np.ndarray: Preprocessed image
        """
        pass

class OCREngine:
    """
    Performs optical character recognition on license plates using Tesseract
    
    TODO:
    - Initialize Tesseract OCR
    - Configure for license plate recognition
    - Parse OCR results
    - Validate plate format
    - Extract alphanumeric characters
    """
    
    def __init__(self, config):
        """
        Initialize OCR engine
        
        Args:
            config (dict): Configuration from config.yaml
        """
        self.config = config
        self.ocr_config = config.get('ocr', {})
        
        self.enabled = self.ocr_config.get('enabled', False)
        self.method = self.ocr_config.get('method', 'tesseract')
        self.min_confidence = self.ocr_config.get('min_confidence', 0.5)
        
        # TODO: Initialize Tesseract if available
        # TODO: Set up Tesseract config (white list characters, etc)
        pass
    
    def recognize_plate(self, plate_image):
        """
        Recognize license plate text using OCR
        
        TODO:
        1. Call Tesseract on plate image
        2. Parse raw OCR output
        3. Extract plate number string
        4. Validate format (should contain alphanumeric)
        5. Filter by confidence threshold
        6. Return recognized plate or None
        
        Args:
            plate_image (np.ndarray): Preprocessed plate image
        
        Returns:
            str: Recognized plate text, or None if unconfident
        """
        pass
    
    def validate_plate_format(self, plate_text):
        """
        Validate if text matches expected plate format
        
        TODO:
        - Check length (typically 6-8 characters)
        - Check character types (letters + numbers)
        - Check for common OCR errors (0 vs O, 1 vs I, etc)
        - Return True if valid format
        
        Args:
            plate_text (str): OCR output
        
        Returns:
            bool: True if valid plate format
        """
        pass
    
    def correct_common_errors(self, plate_text):
        """
        Correct common OCR misrecognitions
        
        TODO:
        - Replace 0 with O if context suggests letter
        - Replace 1 with I if context suggests letter
        - Replace l with 1 if context suggests number
        - Return corrected text
        
        Args:
            plate_text (str): Raw OCR output
        
        Returns:
            str: Corrected text
        """
        pass

class PlateLogger:
    """
    Logs detected license plates with vehicle info
    
    TODO:
    - Store plate detections in database or file
    - Link plate to vehicle ID
    - Link plate to violations/events
    - Query plate history
    """
    
    def __init__(self, config):
        """
        Initialize plate logger
        
        Args:
            config (dict): Configuration from config.yaml
        """
        self.config = config
        # TODO: Initialize database or file logger
        pass
    
    def log_plate_detection(self, vehicle_id, plate_text, confidence, timestamp, frame_number):
        """
        Log detected license plate
        
        TODO:
        - Create record with vehicle ID, plate, confidence, timestamp
        - Store in database or CSV file
        - Link to frame number for later reference
        
        Args:
            vehicle_id (int): Detected vehicle ID
            plate_text (str): Recognized plate text
            confidence (float): OCR confidence (0-1)
            timestamp (float): Detection timestamp
            frame_number (int): Frame number
        """
        pass
    
    def get_plate_history(self, plate_text, time_window=3600):
        """
        Get history of plate detections
        
        TODO:
        - Query all detections of this plate
        - Filter by time window (default 1 hour)
        - Return list of detections with times and locations
        
        Args:
            plate_text (str): License plate
            time_window (int): Seconds to look back
        
        Returns:
            list: Detection records
        """
        pass
