src/signal_control/controller.py
"""
USAD Signal Control Module
Manages traffic light timing and Arduino communication
"""

import serial
import time
import logging
from threading import Thread, Lock

class SignalController:
    """
    Communicates with Arduino to control physical traffic lights
    
    TODO:
    - Initialize serial connection to Arduino
    - Send signal timing commands to Arduino
    - Handle serial communication errors
    - Parse Arduino responses
    - Emergency signal triggering
    """
    
    def __init__(self, config):
        """
        Initialize signal controller with Arduino connection
        
        Args:
            config (dict): Configuration from config.yaml
        """
        self.config = config
        self.arduino_config = config.get('arduino', {})
        
        self.enabled = self.arduino_config.get('enabled', True)
        self.port = self.arduino_config.get('port', 'COM3')
        self.baudrate = self.arduino_config.get('baudrate', 9600)
        self.timeout = self.arduino_config.get('timeout', 2)
        
        self.serial_conn = None
        self.last_command_time = 0
        self.serial_cooldown = self.arduino_config.get('serial_cooldown', 0.1)
        self.command_lock = Lock()
        
        # TODO: Initialize serial connection
        # TODO: Verify connection with Arduino
        pass
    
    def connect(self):
        """
        Establish serial connection to Arduino
        
        TODO:
        - Create serial.Serial connection
        - Wait for Arduino to initialize
        - Read and verify startup message
        - Log connection status
        """
        pass
    
    def send_signal_command(self, direction, green_time, yellow_time, red_time):
        """
        Send signal timing command to Arduino
        
        Format: DIRECTION,GREEN_TIME,YELLOW_TIME,RED_TIME
        Example: N,30,5,40
        
        TODO:
        - Validate timing values
        - Format command string
        - Send via serial with rate limiting
        - Read and log Arduino response
        - Handle errors gracefully
        
        Args:
            direction (str): 'N', 'S', 'E', or 'W'
            green_time (int): Green light duration in seconds
            yellow_time (int): Yellow light duration in seconds
            red_time (int): Red light duration in seconds
        
        Returns:
            bool: True if successful, False otherwise
        """
        pass
    
    def trigger_emergency(self):
        """
        Activate emergency mode on Arduino
        
        TODO:
        - Send "EMERGENCY" command
        - Log emergency activation
        """
        pass
    
    def clear_emergency(self):
        """
        Clear emergency mode on Arduino
        
        TODO:
        - Send "CLEAR_EMERGENCY" command
        - Log emergency clear
        """
        pass
    
    def get_status(self):
        """
        Request current status from Arduino
        
        TODO:
        - Send "STATUS" command
        - Parse Arduino response
        - Return status dictionary
        """
        pass
    
    def read_response(self):
        """
        Read and parse response from Arduino
        
        TODO:
        - Read line from serial
        - Decode and return
        - Handle timeout
        """
        pass
    
    def disconnect(self):
        """
        Close serial connection
        
        TODO:
        - Close serial port
        - Log disconnection
        """
        pass

class SignalOptimizer:
    """
    Calculates optimal signal timing based on vehicle counts and congestion
    
    TODO:
    - Analyze vehicle counts per lane
    - Detect congestion conditions
    - Calculate fairness (ensure all directions get green time)
    - Generate adaptive signal timing
    - Prioritize emergency routes
    """
    
    def __init__(self, config):
        """
        Initialize signal optimizer
        
        Args:
            config (dict): Configuration from config.yaml
        """
        self.config = config
        self.signal_timing_config = config.get('signal_timing', {})
        
        # Base timing
        self.default_green = self.signal_timing_config.get('default_green', 30)
        self.default_yellow = self.signal_timing_config.get('default_yellow', 5)
        self.default_red = self.signal_timing_config.get('default_red', 40)
        
        # Constraints
        self.min_green = self.signal_timing_config.get('min_green', 15)
        self.max_green = self.signal_timing_config.get('max_green', 120)
        self.min_red = self.signal_timing_config.get('min_red', 20)
        self.max_red = self.signal_timing_config.get('max_red', 120)
        
        # Congestion parameters
        self.congestion_threshold = self.signal_timing_config.get('congestion_threshold', 8)
        self.light_congestion_boost = self.signal_timing_config.get('light_congestion_boost', 10)
        self.heavy_congestion_boost = self.signal_timing_config.get('heavy_congestion_boost', 20)
        pass
    
    def optimize_timing(self, lane_vehicles):
        """
        Calculate optimal signal timing based on vehicle counts
        
        TODO:
        1. Analyze vehicle distribution across lanes
        2. Detect congestion in each lane
        3. Assign priority based on:
           - Vehicle count
           - Congestion level
           - Time since last green
        4. Calculate green time for each direction:
           - Base: default_green
           - Add boost if congested
           - Ensure fairness (not too unbalanced)
        5. Calculate corresponding red times
        6. Return signal sequence
        
        Args:
            lane_vehicles (dict): {direction: vehicle_count}
                Example: {'N': 12, 'S': 3, 'E': 8, 'W': 5}
        
        Returns:
            list: Ordered signal commands [(direction, green, yellow, red), ...]
        """
        pass
    
    def detect_congestion(self, vehicle_count):
        """
        Classify congestion level
        
        TODO:
        - Light: above threshold but < heavy threshold
        - Heavy: > heavy threshold
        - Normal: < threshold
        
        Returns:
            str: 'normal', 'light', or 'heavy'
        """
        pass
    
    def calculate_boost(self, congestion_level):
        """
        Calculate green time boost for congestion level
        
        TODO:
        - Return light_congestion_boost for 'light'
        - Return heavy_congestion_boost for 'heavy'
        - Return 0 for 'normal'
        """
        pass
