requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
PyYAML>=6.0
Pillow>=10.0.0
pyserial>=3.5
pytesseract>=0.3.10
twilio>=8.10.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
